requirements.txt
requests==2.26.0
python-dotenv==0.21.0
influxdb-client==1.32.0
